Update AddressBook to ReadyBakey
Updating of all occurrences of AddressBook. Please help me do a second pass to see if there are still any occurrences of it. Noticed there are some Javadoc comments that may or may not contain address book. Hoping that I managed to catch all of them.

Changed the `addressbook.json` to be `readybakey.json` instead. You can delete the `addressbook.json` and it will generate the new `readybakey.json` in the data folder.

Noticed that the AddRemark.md is very outdated. Let's try to see how we can further improve this and keep it updated.